isolate snapshot test (and test setup and teardown)
